requirements.txt
tastytrade==9.13
flask
python-dotenv
pytz
requests
yfinance
pandas
numpy
